fix(deps): update rust crates libsqlite3-sys to 0.36.0 and rusqlite to 0.38.0

+++ src/command/snapshot/set_index.rs
@@ -43,7 +43,7 @@ impl SetIndexCommand {
                 UPDATE metadata
                 SET raft_index = ?
             ",
-            params![index],
+            params![index as i64],
         )?;
         Ok(())
     }
